feat(stalk): single stalk feature
This commit changes the way that the map markers are fetched based on a new prop that allows for the fetching of specific names instead of all markers.

The commit also adds a new navigation feature that allows users to view markers from a specific name and also shows the markers from all names when clicking the "View All" button.